Entity handles that can be default initialized, to show difference vs Resource::Handle.
Ideally handle invalid state would be 0, and a default/invalid resource/entity would be
present at 0. This avoids crashes while giving feedback at runtime that something is
wrong, since the user would observe the default resource and notice it's not intended.

Aliasing is inherently unsafe in the presence of reclammation. Strategies include:
- Don't alias
- Don't reclaim resources (pretend resources are infinite like a garbage collector)
- Reference count aliases (generally too expensive)
- Trap aliasing errors via special handles