Player movement handler cannot be nullptr
Using constructors we can know that on_player_moved can never be null, eliminating
all of the null checks.

Private ctors means make_unique will not work, so resort to new operator.

Change Direction to enum class so it can be forward declared.

on_player_move is injecting Player with logic from Game state, which is questionable.
It does this via a reference in the closure, which happens to be safe because Game
state owns all the game references because it owns the entity controller. Still, this
is a potential aliasing problem in the future.