Fixed bug in Settings introduced during UI refactor
The UI's were changed to build the arrays in their ctors. The arrays sometimes
depend on game state, making this not a great solution. Right now awake() calls
the ctors again (wasteful) as a matter of simplicity. Current solution seems sketchy,
so reverting to how it used to build the UI in awake() might be the better solution.

Consolidated data IO.