Fixed bug in Owner::destroy for resources
Owner didn't originally have a default or invalid state, which leads to problems when
trying to move it. Moves requires the object be in some valid state always, as the dtor
will still be called on it. This bug can happen when operator= is called, destroying once,
and then later the dtor calls for a double free if the given deleter didn't put the object
into an "invalid" or "empty" state. A better solution would be that Owner requires
some indeterminate state that it can compare and dole out deletes itself.